js: Promote core operations to JoinServer
