+++ pkg/joinserver/joinserver.go
@@ -16,9 +16,13 @@
 package joinserver
 
 import (
+	"bytes"
 	"context"
+	"encoding/binary"
 	"io"
+	"math"
 	"math/rand"
+	"sort"
 	"sync"
 	"time"
 
@@ -26,6 +30,9 @@ import (
 	"github.com/oklog/ulid"
 	"go.thethings.network/lorawan-stack/pkg/cluster"
 	"go.thethings.network/lorawan-stack/pkg/component"
+	"go.thethings.network/lorawan-stack/pkg/crypto/cryptoservices"
+	"go.thethings.network/lorawan-stack/pkg/crypto/cryptoutil"
+	"go.thethings.network/lorawan-stack/pkg/encoding/lorawan"
 	"go.thethings.network/lorawan-stack/pkg/interop"
 	"go.thethings.network/lorawan-stack/pkg/log"
 	"go.thethings.network/lorawan-stack/pkg/rpcmiddleware/hooks"
@@ -121,4 +128,316 @@ func (js *JoinServer) RegisterHandlers(s *runtime.ServeMux, conn *grpc.ClientCon
 // RegisterInterop registers the NS-JS and AS-JS interop services.
 func (js *JoinServer) RegisterInterop(srv *interop.Server) {
 	srv.RegisterJS(js.interop)
+}
+
+var supportedMACVersions = [...]ttnpb.MACVersion{
+	ttnpb.MAC_V1_0,
+	ttnpb.MAC_V1_0_1,
+	ttnpb.MAC_V1_0_2,
+	ttnpb.MAC_V1_0_3,
+	ttnpb.MAC_V1_1,
+}
+
+// HandleJoin handles the given join-request.
+func (js *JoinServer) HandleJoin(ctx context.Context, req *ttnpb.JoinRequest) (res *ttnpb.JoinResponse, err error) {
+	logger := log.FromContext(ctx)
+	defer func() {
+		if err != nil {
+			registerRejectJoin(ctx, req, err)
+		}
+	}()
+
+	supported := false
+	for _, ver := range supportedMACVersions {
+		if req.SelectedMACVersion == ver {
+			supported = true
+			break
+		}
+	}
+	if !supported {
+		return nil, errUnsupportedLoRaWANVersion.WithAttributes("version", req.SelectedMACVersion)
+	}
+
+	req.Payload = &ttnpb.Message{}
+	if err = lorawan.UnmarshalMessage(req.RawPayload, req.Payload); err != nil {
+		return nil, errDecodePayload.WithCause(err)
+	}
+
+	if req.Payload.Major != ttnpb.Major_LORAWAN_R1 {
+		return nil, errUnsupportedLoRaWANVersion.WithAttributes("version", req.Payload.Major)
+	}
+	if req.Payload.MType != ttnpb.MType_JOIN_REQUEST {
+		return nil, errWrongPayloadType.WithAttributes("type", req.Payload.MType)
+	}
+
+	pld := req.Payload.GetJoinRequestPayload()
+	if pld == nil {
+		return nil, errNoJoinRequest
+	}
+	if pld.DevEUI.IsZero() {
+		return nil, errNoDevEUI
+	}
+	if pld.JoinEUI.IsZero() {
+		return nil, errNoJoinEUI
+	}
+
+	match := false
+	for _, p := range js.euiPrefixes {
+		if p.Matches(pld.JoinEUI) {
+			match = true
+			break
+		}
+	}
+	switch {
+	case !match && req.SelectedMACVersion.Compare(ttnpb.MAC_V1_1) < 0:
+		return nil, errUnknownAppEUI
+	case !match:
+		// TODO: Determine the cluster containing the device.
+		// https://github.com/TheThingsNetwork/lorawan-stack/issues/4
+		return nil, errForwardJoinRequest
+	}
+
+	dev, err := js.devices.SetByEUI(ctx, pld.JoinEUI, pld.DevEUI,
+		[]string{
+			"last_dev_nonce",
+			"last_join_nonce",
+			"resets_join_nonces",
+			"root_keys",
+			"used_dev_nonces",
+			"provisioner_id",
+			"provisioning_data",
+		},
+		func(dev *ttnpb.EndDevice) (*ttnpb.EndDevice, []string, error) {
+			paths := make([]string, 0, 3)
+
+			dn := uint32(binary.BigEndian.Uint16(pld.DevNonce[:]))
+			if req.SelectedMACVersion.Compare(ttnpb.MAC_V1_1) >= 0 {
+				if (dn != 0 || dev.LastDevNonce != 0 || dev.LastJoinNonce != 0) && !dev.ResetsJoinNonces {
+					if dn <= dev.LastDevNonce {
+						return nil, nil, errDevNonceTooSmall
+					}
+					if dn == math.MaxUint32 {
+						return nil, nil, errDevNonceTooHigh
+					}
+				}
+				dev.LastDevNonce = dn
+				paths = append(paths, "last_dev_nonce")
+			} else {
+				i := sort.Search(len(dev.UsedDevNonces), func(i int) bool { return dev.UsedDevNonces[i] >= dn })
+				if i < len(dev.UsedDevNonces) && dev.UsedDevNonces[i] == dn {
+					return nil, nil, errReuseDevNonce
+				}
+				dev.UsedDevNonces = append(dev.UsedDevNonces, 0)
+				copy(dev.UsedDevNonces[i+1:], dev.UsedDevNonces[i:])
+				dev.UsedDevNonces[i] = dn
+				paths = append(paths, "used_dev_nonces")
+			}
+
+			var b []byte
+			if req.CFList == nil {
+				b = make([]byte, 0, 17)
+			} else {
+				b = make([]byte, 0, 33)
+			}
+			b, err = lorawan.AppendMHDR(b, ttnpb.MHDR{
+				MType: ttnpb.MType_JOIN_ACCEPT,
+				Major: req.Payload.Major,
+			})
+			if err != nil {
+				return nil, nil, errEncodePayload.WithCause(err)
+			}
+
+			if dev.LastJoinNonce >= 1<<24-1 {
+				return nil, nil, errJoinNonceTooHigh
+			}
+			dev.LastJoinNonce++
+			paths = append(paths, "last_join_nonce")
+
+			var jn types.JoinNonce
+			nb := make([]byte, 4)
+			binary.BigEndian.PutUint32(nb, dev.LastJoinNonce)
+			copy(jn[:], nb[1:])
+
+			b, err = lorawan.AppendJoinAcceptPayload(b, ttnpb.JoinAcceptPayload{
+				NetID:      req.NetID,
+				JoinNonce:  jn,
+				CFList:     req.CFList,
+				DevAddr:    req.DevAddr,
+				DLSettings: req.DownlinkSettings,
+				RxDelay:    req.RxDelay,
+			})
+			if err != nil {
+				return nil, nil, errEncodePayload.WithCause(err)
+			}
+
+			js.entropyMu.Lock()
+			skID, err := ulid.New(ulid.Timestamp(time.Now()), js.entropy)
+			js.entropyMu.Unlock()
+			if err != nil {
+				return nil, nil, errGenerateSessionKeyID
+			}
+
+			cs := js.GetPeer(ctx, ttnpb.PeerInfo_CRYPTO_SERVER, dev.EndDeviceIdentifiers)
+
+			var networkCryptoService cryptoservices.Network
+			if req.SelectedMACVersion.Compare(ttnpb.MAC_V1_1) >= 0 && dev.RootKeys != nil && dev.RootKeys.NwkKey != nil {
+				// LoRaWAN 1.1 and higher use a NwkKey.
+				nwkKey, err := cryptoutil.UnwrapAES128Key(*dev.RootKeys.NwkKey, js.KeyVault)
+				if err != nil {
+					return nil, nil, err
+				}
+				networkCryptoService = cryptoservices.NewMemory(&nwkKey, nil)
+			} else if cs != nil {
+				networkCryptoService = cryptoservices.NewNetworkRPCClient(cs.Conn(), js.KeyVault, js.WithClusterAuth())
+			}
+
+			var applicationCryptoService cryptoservices.Application
+			if dev.RootKeys != nil && dev.RootKeys.AppKey != nil {
+				appKey, err := cryptoutil.UnwrapAES128Key(*dev.RootKeys.AppKey, js.KeyVault)
+				if err != nil {
+					return nil, nil, err
+				}
+				applicationCryptoService = cryptoservices.NewMemory(nil, &appKey)
+				if req.SelectedMACVersion.Compare(ttnpb.MAC_V1_1) < 0 {
+					// LoRaWAN 1.0.x use the AppKey for network security operations.
+					networkCryptoService = cryptoservices.NewMemory(nil, &appKey)
+				}
+			} else if cs != nil {
+				applicationCryptoService = cryptoservices.NewApplicationRPCClient(cs.Conn(), js.KeyVault, js.WithClusterAuth())
+			}
+			if networkCryptoService == nil {
+				return nil, nil, errNoNwkKey
+			}
+			if applicationCryptoService == nil {
+				return nil, nil, errNoAppKey
+			}
+
+			cryptoDev := &ttnpb.EndDevice{}
+			if err := cryptoDev.SetFields(dev, "ids", "provisioner_id", "provisioning_data"); err != nil {
+				return nil, nil, err
+			}
+			reqMIC, err := networkCryptoService.JoinRequestMIC(ctx, cryptoDev, req.SelectedMACVersion, req.RawPayload[:19])
+			if err != nil {
+				return nil, nil, errComputeMIC.WithCause(err)
+			}
+			if !bytes.Equal(reqMIC[:], req.RawPayload[19:]) {
+				return nil, nil, errMICMismatch
+			}
+			resMIC, err := networkCryptoService.JoinAcceptMIC(ctx, cryptoDev, req.SelectedMACVersion, 0xff, pld.DevNonce, b)
+			if err != nil {
+				return nil, nil, errComputeMIC.WithCause(err)
+			}
+			enc, err := networkCryptoService.EncryptJoinAccept(ctx, cryptoDev, req.SelectedMACVersion, append(b[1:], resMIC[:]...))
+			if err != nil {
+				return nil, nil, errEncryptPayload.WithCause(err)
+			}
+			nwkSKeys, err := networkCryptoService.DeriveNwkSKeys(ctx, cryptoDev, req.SelectedMACVersion, jn, pld.DevNonce, req.NetID)
+			if err != nil {
+				return nil, nil, errDeriveNwkSKeys.WithCause(err)
+			}
+			appSKey, err := applicationCryptoService.DeriveAppSKey(ctx, cryptoDev, req.SelectedMACVersion, jn, pld.DevNonce, req.NetID)
+			if err != nil {
+				return nil, nil, errDeriveAppSKey.WithCause(err)
+			}
+
+			sessionKeys := ttnpb.SessionKeys{
+				SessionKeyID: skID[:],
+				FNwkSIntKey: &ttnpb.KeyEnvelope{
+					// TODO: Encrypt key with NS KEK https://github.com/TheThingsNetwork/lorawan-stack/issues/5
+					Key: &nwkSKeys.FNwkSIntKey,
+				},
+				AppSKey: &ttnpb.KeyEnvelope{
+					// TODO: Encrypt key with AS KEK https://github.com/TheThingsNetwork/lorawan-stack/issues/5
+					Key: &appSKey,
+				},
+			}
+			if req.SelectedMACVersion == ttnpb.MAC_V1_1 {
+				sessionKeys.SNwkSIntKey = &ttnpb.KeyEnvelope{
+					// TODO: Encrypt key with NS KEK https://github.com/TheThingsNetwork/lorawan-stack/issues/5
+					Key: &nwkSKeys.SNwkSIntKey,
+				}
+				sessionKeys.NwkSEncKey = &ttnpb.KeyEnvelope{
+					// TODO: Encrypt key with NS KEK https://github.com/TheThingsNetwork/lorawan-stack/issues/5
+					Key: &nwkSKeys.NwkSEncKey,
+				}
+			}
+
+			res = &ttnpb.JoinResponse{
+				RawPayload:  append(b[:1], enc...),
+				SessionKeys: sessionKeys,
+			}
+			_, err = js.keys.SetByID(ctx, *dev.DevEUI, res.SessionKeys.SessionKeyID,
+				[]string{
+					"session_key_id",
+					"f_nwk_s_int_key",
+					"s_nwk_s_int_key",
+					"nwk_s_enc_key",
+					"app_s_key",
+				},
+				func(stored *ttnpb.SessionKeys) (*ttnpb.SessionKeys, []string, error) {
+					if stored != nil {
+						return nil, nil, errDuplicateIdentifiers
+					}
+					return &res.SessionKeys, []string{
+						"session_key_id",
+						"f_nwk_s_int_key",
+						"s_nwk_s_int_key",
+						"nwk_s_enc_key",
+						"app_s_key",
+					}, nil
+				},
+			)
+			if err != nil {
+				return nil, nil, err
+			}
+
+			dev.Session = &ttnpb.Session{
+				StartedAt:   time.Now().UTC(),
+				DevAddr:     req.DevAddr,
+				SessionKeys: res.SessionKeys,
+			}
+			paths = append(paths, "session")
+
+			return dev, paths, nil
+		})
+	if err != nil {
+		logger.WithFields(log.Fields(
+			"join_eui", pld.JoinEUI,
+			"dev_eui", pld.DevEUI,
+		)).WithError(err).Error("Failed to update device")
+		return nil, err
+	}
+
+	registerAcceptJoin(ctx, dev, req)
+	return res, nil
+}
+
+// GetNwkSKeys returns the requested network session keys.
+func (js *JoinServer) GetNwkSKeys(ctx context.Context, req *ttnpb.SessionKeyRequest) (*ttnpb.NwkSKeysResponse, error) {
+	ks, err := js.keys.GetByID(ctx, req.DevEUI, req.SessionKeyID,
+		[]string{
+			"f_nwk_s_int_key",
+			"nwk_s_enc_key",
+			"s_nwk_s_int_key",
+		},
+	)
+	if err != nil {
+		return nil, errRegistryOperation.WithCause(err)
+	}
+
+	if ks.NwkSEncKey == nil {
+		return nil, errNoNwkSEncKey
+	}
+	if ks.FNwkSIntKey == nil {
+		return nil, errNoFNwkSIntKey
+	}
+	if ks.SNwkSIntKey == nil {
+		return nil, errNoSNwkSIntKey
+	}
+
+	return &ttnpb.NwkSKeysResponse{
+		NwkSEncKey:  *ks.NwkSEncKey,
+		FNwkSIntKey: *ks.FNwkSIntKey,
+		SNwkSIntKey: *ks.SNwkSIntKey,
+	}, nil
 }
